+++ src/pages/Collect/Collect.tsx
@@ -1,9 +1,9 @@
-import { useLocalStorage } from '@react-query/hooks/useLocalStorage'
+import { useStorage } from '@react-query/hooks/useStorage'
 import React from 'react'
 import RecentRecord from './RecentRecord'
 
 export default function Collect() {
-  useLocalStorage()
+  useStorage('sessionStorage')
   return (
     <div className="relative flex flex-col ">
       <div className="h-[1000px] bg-primary-6">1</div>
