+++ src/pages/Main/Main.tsx
@@ -1,6 +1,6 @@
 import { CELEBRATION_ID } from '@assets/constant/constant'
 import ParentCategoryTab from '@components/ParrentCategoryTab'
-import { useLocalStorage } from '@react-query/hooks/useLocalStorage'
+import { useStorage } from '@react-query/hooks/useStorage'
 import React, { useState } from 'react'
 import { parentCategoryID } from 'types/category'
 import MixRecord from './MixRecord'
@@ -11,7 +11,7 @@ export default function Main() {
   const [parentCategoryId, setParentCategoryId] =
     useState<parentCategoryID>(CELEBRATION_ID)
 
-  useLocalStorage()
+  useStorage('sessionStorage')
   return (
     <>
       <section id="mixRecord">
